Refactored server
